# doc/api.yaml
openapi: 3.0.0
info:
  title: Knowledge Hub
  description: Knowledge hub service for managing articles, categories, and comments
  version: 1.0.0
servers:
  - url: /
components:
  schemas:
...
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
          example: 'Great article!'
        articleId:
          type: string
          format: uuid
        authorId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: number
          example: 1655000000
      required:
        - id
        - content
        - articleId
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
    ForbiddenError:
      description: Insufficient permissions for this operation
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /comment:
    get:
      tags:
        - Comment
      summary: Get comments for an article
      description: Gets all comments for a specific article. Requires articleId query parameter.
      parameters:
        - name: articleId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Article ID to get comments for
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        401:
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Comment
      summary: Add new comment
      description: Add a comment to an article (editor can create own, admin can create any)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                articleId:
                  type: string
                  format: uuid
                authorId:
                  type: string
                  format: uuid
                  nullable: true
              required:
                - content
                - articleId
      responses:
        201:
          description: Comment is created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        400:
          description: Bad request. body does not contain required fields
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        422:
          description: Article with given articleId doesn't exist
  /comment/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Comment
      summary: Get single comment by id
      description: Gets single comment by id
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        400:
          description: Bad request. commentId is invalid (not uuid)
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Comment was not found
    delete:
      tags:
        - Comment
      summary: Delete comment
      description: Delete comment (admin can delete any, editor can delete own)
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request. commentId is invalid (not uuid)
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          description: Comment was not found